move CareerImgEdit to component, chg import path
