+++ client/src/views/in/recruitment/RecruitmentForm.js
@@ -3,7 +3,7 @@ import React, { useState, useRef } from 'react'
 import { useSelector, useDispatch } from 'react-redux'
 import { selectCareer, clearCroppedDataUrl, clearCroppedFile } from '../../../slices/careerSlice'
 import { useHistory } from 'react-router'
-import CareerImageEditor from '../career/CareerImageEditor'
+import CareerImageEditor from '../../components/CareerImageEditor'
 import JoditEditor from 'jodit-react'
 import ReactTooltip from 'react-tooltip'
 import PropTypes from 'prop-types'
